+++ app/components/channel-invitation-confirm.html
