+++ app/components/channel-member-update.html
