+++ app/components/index.ts
@@ -4,9 +4,11 @@ import authReturn from './auth-return';
 import avatar from './avatar';
 import avatarList from './avatar-list';
 import channel from './channel';
+import channelInvitationConfirm from './channel-invitation-confirm';
 import channelInvitationNew from './channel-invitation-new';
 import channelInvitations from './channel-invitations';
 import channelMembers from './channel-members';
+import channelMemberUpdate from './channel-member-update';
 import channelSettings from './channel-settings';
 import channels from './channels';
 import channelsInvitations from './channels-invitations';
@@ -33,7 +35,6 @@ import hivedDocs from './hived-docs';
 import hivers from './hivers';
 import jobs from './jobs';
 import inlineEditable from './inline-editable';
-import invitationConfirm from './invitation-confirm';
 import legalNotice from './legal-notice';
 import login from './login';
 import mainPage from './main-page';
@@ -44,7 +45,6 @@ import marginDiscussionEdit from './margin-discussion-edit';
 import marginDiscussionPlaceholder from './margin-discussion-placeholder';
 import marginDiscussions from './margin-discussions';
 import marginReplyEdit from './margin-reply-edit';
-import memberUpdate from './member-update';
 import navbar from './navbar';
 import navbarSearch from './navbar-search';
 import navbarUser from './navbar-user';
@@ -82,6 +82,7 @@ export default function(app) {
   avatar(app);
   avatarList(app);
   channel(app);
+  channelInvitationConfirm(app);
   channelInvitationNew(app);
   channelInvitations(app);
   channelMembers(app);
@@ -109,7 +110,6 @@ export default function(app) {
   hiveButton(app);
   hivedDocs(app);
   hivers(app);
-  invitationConfirm(app);
   jobs(app);
   inlineEditable(app);
   legalNotice(app);
